// src/MarketSale/MarketSale.concrete.typeInfo.d.ts
export interface OtherSaleState {
    progressDetails: /*minStructField*/ SaleProgressDetails
    salePace: /*minStructField*/ number
    state: /*minStructField*/ MarketSaleState
}
// ...
export interface OtherSaleStateLike {
    progressDetails: /*minStructField*/ SaleProgressDetailsLike
    salePace: /*minStructField*/ number
    state: /*minStructField*/ MarketSaleStateLike
}
// ...
export interface FixedSaleDetails {
    settings: /*minStructField*/ DynamicSaleV1Settings
    startAt: /*minStructField*/ number
    vxfTokensTo: /*minStructField*/ VxfDestination | undefined
    vxfFundsTo: /*minStructField*/ VxfDestination | undefined
}
// ...
export interface FixedSaleDetailsLike {
    settings: /*minStructField*/ DynamicSaleV1SettingsLike
    startAt: /*minStructField*/ TimeLike
    vxfTokensTo: /*minStructField*/ VxfDestinationLike | undefined
    vxfFundsTo: /*minStructField*/ VxfDestinationLike | undefined
}



/**
 * A strong type for the canonical form of SaleAssets
 * @remarks
 * Note that any enum fields in this type are expressed as a disjoint union of the enum variants.  Processing
 * enum data conforming to this type can be a bit of a pain.
 * For a more ergonomic, though less strictly-safe form of this type, see ErgoSaleAssets instead.
 * @public
 */
export interface SaleAssets {
    saleUnitAssets: /*minStructField*/ Value
    singleBuyMaxUnits: /*minStructField*/ bigint
    primaryAssetMph: /*minStructField*/ MintingPolicyHash
    primaryAssetName: /*minStructField*/ number[]
    primaryAssetTargetCount: /*minStructField*/ bigint
    totalSaleUnits: /*minStructField*/ bigint
}
// ...
export interface SaleAssetsLike {
    saleUnitAssets: /*minStructField*/ Value | [MintingPolicyHash | string | number[], [number[] | string, IntLike][]][] | {mph: MintingPolicyHash | string | number[], tokens: {name: number[] | string, qty: IntLike}[]}[]
    singleBuyMaxUnits: /*minStructField*/ IntLike
    primaryAssetMph: /*minStructField*/ MintingPolicyHash | string | number[]
    primaryAssetName: /*minStructField*/ number[]
    primaryAssetTargetCount: /*minStructField*/ IntLike
    totalSaleUnits: /*minStructField*/ IntLike
}



/**
 * A strong type for the canonical form of ThreadInfo
 * @remarks
 * Note that any enum fields in this type are expressed as a disjoint union of the enum variants.  Processing
 * enum data conforming to this type can be a bit of a pain.
 * For a more ergonomic, though less strictly-safe form of this type, see ErgoThreadInfo instead.
 * @public
 */
export interface ThreadInfo {
    nestedThreads: /*minStructField*/ bigint
    retiredThreads: /*minStructField*/ bigint
    parentChunkId: /*minStructField*/ number[]
    chunkForkedAt: /*minStructField*/ number
    saleId: /*minStructField*/ number[]
}
// ...
export interface ThreadInfoLike {
    nestedThreads: /*minStructField*/ IntLike
    retiredThreads: /*minStructField*/ IntLike
    parentChunkId: /*minStructField*/ number[]
    chunkForkedAt: /*minStructField*/ TimeLike
    saleId: /*minStructField*/ number[]
}



/**
 * A strong type for the canonical form of MoreFields
 * @remarks
 * Note that any enum fields in this type are expressed as a disjoint union of the enum variants.  Processing
 * enum data conforming to this type can be a bit of a pain.
 * For a more ergonomic, though less strictly-safe form of this type, see ErgoMoreFields instead.
 * @public
 */
export interface MoreFields {
    saleState: /*minStructField*/ OtherSaleState
    fixedSaleDetails: /*minStructField*/ FixedSaleDetails
    saleAssets: /*minStructField*/ SaleAssets
    threadInfo: /*minStructField*/ ThreadInfo
}
// ...
export interface MoreFieldsLike {
    saleState: /*minStructField*/ OtherSaleStateLike
    fixedSaleDetails: /*minStructField*/ FixedSaleDetailsLike
    saleAssets: /*minStructField*/ SaleAssetsLike
    threadInfo: /*minStructField*/ ThreadInfoLike
}



/**
 * A strong type for the canonical form of MarketSaleData
 * @remarks
 * Note that any enum fields in this type are expressed as a disjoint union of the enum variants.  Processing
 * enum data conforming to this type can be a bit of a pain.
 * For a more ergonomic, though less strictly-safe form of this type, see ErgoMarketSaleData instead.
 * @public
 */
export interface MarketSaleData {
    id: /*minStructField*/ number[]
    type: /*minStructField*/ string
    name: /*minStructField*/ string
    moreFields: /*minStructField*/ MoreFields
}
// ...
export interface MarketSaleDataLike {
    id: /*minStructField*/ number[]
    type: /*minStructField*/ string
    name: /*minStructField*/ string
    moreFields: /*minStructField*/ MoreFieldsLike
}


/**
 * expresses the essential fields needed for initiating creation of a MarketSaleData
 * @public
 */
export type minimalMarketSaleData = minimalData<MarketSaleDataLike>